backend/requirements.txt
Django==5.2
djangorestframework==3.16.0
markdown==3.7
python-dotenv==1.0.0
python-frontmatter==1.0.0
PyYAML==6.0.1
python-multipart==0.0.6
fastapi==0.109.0
uvicorn==0.27.0
